[BUGFIX] Fix integer overflow in RingBuffer indices.
Normalize indexInput and indexOutput when they are both higher than the
full buffer size.
This prevents an integer overflow from occurring, which could lead to a
buffer overflow if the cell size is not a power of two.
This problem would typically appear if a large amount of data (> 4GB)
were to be sent.